Add text and confirm to command

+++ src/Commands/CloudflareIpInstall.php
@@ -37,14 +37,28 @@ class CloudflareIpInstall extends Command
      */
     public function handle()
     {
-        copy(__DIR__.'/../cloudflare-ip.php', base_path('bootstrap/cloudflare-ip.php'));
-        copy(base_path('public/index.php'), base_path('public/index.php.backup'));
+        $this->line('Warning! This will modify your index.php file.');
+        $this->line('The following line will be added...');
+        $this->info("require_once __DIR__.'/../bootstrap/cloudflare-ip.php';");
 
-        $file = explode(PHP_EOL, file_get_contents(base_path('public/index.php')));
-        $line = array_search("\$app = require_once __DIR__.'/../bootstrap/app.php';", $file);
-        $firstHalf = array_slice($file, 0, $line+1);
-        array_push($firstHalf, '');
-        array_push($firstHalf, "require_once __DIR__.'/../bootstrap/cloudflare-ip.php';");
-        file_put_contents(base_path('public/index.php'), implode(PHP_EOL, array_merge($firstHalf, array_slice($file, $line+1))));
+        if($this->confirm('Are you sure you want to continue? (yes/no)')) {
+
+            $this->info('Copying cloudflare-ip.php to bootstrap...');
+            copy(__DIR__.'/../cloudflare-ip.php', base_path('bootstrap/cloudflare-ip.php'));
+
+            $this->info('Backing up index.php...');
+            copy(base_path('public/index.php'), base_path('public/index.php.backup'));
+
+            $this->info('Adding line to index.php...');
+            $file = explode(PHP_EOL, file_get_contents(base_path('public/index.php')));
+            $line = array_search("\$app = require_once __DIR__.'/../bootstrap/app.php';", $file);
+            $firstHalf = array_slice($file, 0, $line+1);
+            array_push($firstHalf, '');
+            array_push($firstHalf, "require_once __DIR__.'/../bootstrap/cloudflare-ip.php';");
+            file_put_contents(base_path('public/index.php'), implode(PHP_EOL, array_merge($firstHalf, array_slice($file, $line+1))));
+
+            $this->info('Done.');
+        }
+        
     }
 }
